refactor : 비동기/동기 호출 메서드 별 예외 처리 방식 변경 #58
